+++ CHANGELOG.md
@@ -33,6 +33,7 @@ https://mediaelch.github.io/mediaelch-doc/faq.html#where-are-mediaelchs-settings
  - Implement ScraperSettings class for better mocking
  - Use `QVector<T>` instead of `QList<T>` as the default container
  - Reorganize project structure
+ - Add CMake build system (#700)
 
 
 ## 2.6.0 - Ferenginar (2019-01-06)